+++ clients/include/utility.h
@@ -514,7 +514,8 @@ typedef enum hipblasClientProcessor
     gfx1152 = 1152,
     gfx1153 = 1153,
     gfx1200 = 1200,
-    gfx1201 = 1201
+    gfx1201 = 1201,
+    gfx1250 = 1250
 } hipblasClientProcessor;
 
 /* get architecture number */