+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Documentation for hipBLAS is available at
  
 ### Added
 
-* gfx90c support to clients.
+* gfx1250 and gfx90c support to clients.
 * Version and other properties to Windows `hipblas.dll`.
 * Support for `OpenBLAS` ILP64-based API usage in clients.